Refactoring and Commenting
Changed a few class names and file names, commented many of the methods.
