+++ 7-Logical-Agents/AndExpression.ts
@@ -1,0 +1,6 @@
+export class AndExpression {
+
+  public readonly x = 0;
+  public readonly y = 0;
+
+}
