+++ 7-Logical-Agents/AtomicExpression.ts
@@ -1,0 +1,6 @@
+export class AtomicExpression {
+
+  public readonly x = 0;
+  public readonly y = 0;
+
+}
