+++ 7-Logical-Agents/NotExpression.ts
@@ -1,0 +1,6 @@
+export class NotExpression {
+
+  public readonly x = 0;
+  public readonly y = 0;
+
+}
