+++ 7-Logical-Agents/OrExpression.ts
@@ -1,0 +1,6 @@
+export class OrExpression {
+
+  public readonly x = 0;
+  public readonly y = 0;
+
+}
