+++ 7-Logical-Agents/index.html
@@ -42,14 +42,14 @@
                 <tr>
                     <th class="text-center" scope="row">Red</th>
                     <td style="background-color: #ff0000;">        </td>
-                    <td>Breeze</td>
-                    <td>There is a Pit in one of the Adjacent cells</td>
+                    <td>Stench</td>
+                    <td>Wumpus is close by, adjacent to your cell</td>
                 </tr>
                 <tr>
                     <th class="text-center" scope="row">Black</th>
                     <td style="background-color: #000000;">        </td>
-                    <td>Stench</td>
-                    <td>Wumpus is close by, adjacent to your cell</td>
+                    <td>Breeze</td>
+                    <td>There is a Pit in one of the Adjacent cells</td>
                 </tr>
                 <tr>
                     <th class="text-center" scope="row">Brown</th>
@@ -74,13 +74,13 @@
                 </button>
                 <div style="border: solid black; margin-top: 50px; padding: 20px;">
                     <h3>
-                        <div style="margin-bottom: 20px"><strong>The Agent:</strong></div>
+                        <div style="margin-bottom: 20px"><strong>The UserAgent:</strong></div>
                         <div id="agent-console"><span>I feel safe here, nothing around.</span></div>
                     </h3>
                 </div>
             </div>
             <div class="col-lg-12">
-                <h3>Model Filtering</h3>
+                <h3>Model ModelFiltering</h3>
                 <p style="padding-left: 30px; padding-right: 30px;">
                 When going through the grid, the agent develops models of the world, of where the dangers might lie
                 and tries to find the safe moves by contradicting these models using its knowledge base, which it
