+++ 7-Logical-Agents/main.ts
@@ -1,11 +1,9 @@
-import { Move } from "./agent";
-import { Filtering } from "./filtering";
-import { GodSight } from "./god";
-import { Grid } from "./grid";
+import { GameGrid } from "./GameGrid";
+import { Move } from "./UserAgent";
 
 declare var $: any;
 
-const game: Grid = new Grid();
+const game: GameGrid = new GameGrid();
 game.getTile(3, 1).hasPit = true;
 game.getTile(3, 3).hasPit = true;
 game.getTile(4, 4).hasPit = true;
@@ -13,9 +11,9 @@ game.getTile(1, 3).hasWumpus = true;
 game.getTile(2, 3).hasGold = true;
 game.sensorUpdate();
 
-const god: GodSight = new GodSight(game);
-$("#mode-game").on("click", () => { god.hide(); });
-$("#mode-god").on("click", () => { god.render(); });
+// Binding the click events
+$("#mode-game").on("click", () => { game.godSight.hide(); });
+$("#mode-god").on("click", () => { game.godSight.render(); });
 
 // Binding the Keypress Event
 $("html").on("keydown", (e: any) => {
